Refactor localEventBase.ts
