+++ _common/gameModel/libs/ClientScript/game/common/impacts/implementation/radioImpact.ts
@@ -1,11 +1,11 @@
 import {
-  AddRadioMessageLocalEvent,
   LocalEventBase,
   SourceType,
 } from '../../localEvents/localEventBase';
 import { RadioType } from '../../radio/communicationType';
 import { MainSimulationState } from '../../simulationState/mainSimulationState';
 import { ImpactBase } from '../impact';
+import { AddRadioMessageLocalEvent } from '../../localEvents/localEventBaseRadio';
 
 export interface RadioMessageImpact extends ImpactBase {
   type: 'radio';
