+++ _common/gameModel/libs/ClientScript/game/common/localEvents/localEventBase.ts
@@ -1,63 +1,29 @@
-import {
-  registerHideInactiveActorWarning,
-  registerOpenSelectedActorPanelAfterMove,
-} from '../../../gameInterface/afterUpdateCallbacks';
 import { entries, keys } from '../../../tools/helper';
-import { activableLogger, mainSimLogger, resourceLogger } from '../../../tools/logger';
+import { mainSimLogger, resourceLogger } from '../../../tools/logger';
 import { getTranslation } from '../../../tools/translation';
-import {
-  getCachedHospitalsByProximity,
-  getCachedPatientUnitById,
-  getCachedPatientUnitIdsSorted,
-} from '../../loaders/hospitalLoader';
-import { ActionBase, OnTheRoadAction } from '../actions/actionBase';
-import { Actor, InterventionRole } from '../actors/actor';
-import { getCasuActorId, getHighestAuthorityActorsByLocation } from '../actors/actorLogic';
-import {
-  ActionId,
-  ActorId,
-  GlobalEventId,
-  PatientUnitId,
-  ResourceContainerDefinitionId,
-  ResourceId,
-  SimDuration,
-  SimTime,
-  TaskId,
-  TranslationKey,
-} from '../baseTypes';
-import { FailedRessourceArrivalDelay, TimeSliceDuration } from '../constants';
-import {
-  CasuMessagePayload,
-  HospitalRequestPayload,
-  MethaneMessagePayload,
-} from '../events/casuMessageEvent';
+import { ActionBase } from '../actions/actionBase';
+import { getHighestAuthorityActorsByLocation } from '../actors/actorLogic';
+import { ActorId, GlobalEventId, ResourceContainerDefinitionId, SimDuration, SimTime, TaskId } from '../baseTypes';
+import { FailedRessourceArrivalDelay } from '../constants';
+import { CasuMessagePayload, MethaneMessagePayload } from '../events/casuMessageEvent';
 import { GameOptions } from '../gameOptions';
-import { ActivationOperator } from '../impacts/implementation/activationImpact';
-import { Uid } from '../interfaces';
-import { BuildStatus } from '../mapEntities/mapEntityDescriptor';
-import { computeNewPatientsState } from '../patients/handleState';
-import { formatStandardPretriageReport } from '../patients/pretriageUtils';
 import { RadioType } from '../radio/communicationType';
 import * as RadioLogic from '../radio/radioLogic';
 import { getContainerDef, resolveResourceRequest } from '../resources/emergencyDepartment';
-import { Resource } from '../resources/resource';
 import { ResourceContainerType } from '../resources/resourceContainer';
 import * as ResourceLogic from '../resources/resourceLogic';
 import { resourceArrivalLocationResolution } from '../resources/resourceLogic';
 import { ResourceType } from '../resources/resourceType';
-import { Activable, ChoiceActivable, getChoiceActivable } from '../simulationState/activableState';
-import { updateHospitalProximityRequest } from '../simulationState/hospitalState';
-import { canMoveToLocation, LOCATION_ENUM } from '../simulationState/locationState';
+import { LOCATION_ENUM } from '../simulationState/locationState';
 import { MainSimulationState } from '../simulationState/mainSimulationState';
 import { changePatientLocation, PatientLocation } from '../simulationState/patientState';
 import * as ResourceState from '../simulationState/resourceStateAccess';
 import * as TaskState from '../simulationState/taskStateAccess';
-import { getTaskByTypeAndLocation, getTaskCurrentStatus } from '../simulationState/taskStateAccess';
-import { isTimeForwardReady, updateCurrentTimeFrame } from '../simulationState/timeState';
-import { TaskBase, TaskStatus, TaskType } from '../tasks/taskBase';
-import { getIdleTaskUid } from '../tasks/taskLogic';
+import { TaskBase, TaskStatus } from '../tasks/taskBase';
 import { evaluateAllTriggers, Trigger } from '../triggers/trigger';
 import { getLocalEventManager } from './localEventManager';
+import { AddActorLocalEvent } from './localEventBaseActors';
+import { AddMessageLocalEvent, AddNotificationLocalEvent, AddRadioMessageLocalEvent } from './localEventBaseRadio';
 
 export interface LocalEvent {
   type: string;
@@ -172,258 +138,6 @@ export class PlanActionLocalEvent extends LocalEventBase {
     so.actions.push(this.props.action);
     // init action
     this.props.action.update(state);
-  }
-}
-
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-// time
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-
-export abstract class TimeForwardLocalBaseEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly priority?: number;
-      readonly type: string;
-      readonly actors: ActorId[];
-    }
-  ) {
-    const defaultProps = { priority: 1 };
-    super({ ...defaultProps, ...props });
-  }
-
-  protected updateCurrentTimeFrame(state: MainSimulationState, modifier: number) {
-    updateCurrentTimeFrame(state, this.props.actors, modifier, this.props.simTimeStamp);
-  }
-}
-
-/**
- * When applied to state, bumps the readiness of the provided actors.
- * If all actors are ready, time forwards
- */
-export class TimeForwardLocalEvent extends TimeForwardLocalBaseEvent {
-  constructor(
-    readonly extensionProps: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly actors: ActorId[];
-      readonly timeJump: number;
-    }
-  ) {
-    super({ ...extensionProps, type: 'TimeForwardLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    this.updateCurrentTimeFrame(state, 1);
-    if (isTimeForwardReady(state)) {
-      state.incrementSimulationTime(this.extensionProps.timeJump);
-
-      // update patients
-      this.updatePatients(state, this.extensionProps.timeJump);
-
-      // update all actions
-      this.updateActions(state);
-
-      // update all tasks
-      this.updateTasks(state);
-
-      // run the triggers
-      const generatedLocalEvents = evaluateAllTriggers(state);
-      getLocalEventManager().queueLocalEvents(generatedLocalEvents);
-
-      registerOpenSelectedActorPanelAfterMove();
-      registerHideInactiveActorWarning();
-
-      state.updateForwardTimeFrame();
-
-      // auto-continue if all actors are still awaiting
-      if (isTimeForwardReady(state)) {
-        const tfw = new TimeForwardLocalEvent({
-          parentEventId: this.extensionProps.parentEventId,
-          source: this.extensionProps.source,
-          simTimeStamp: state.getSimTime(),
-          actors: [],
-          timeJump: TimeSliceDuration,
-        });
-        getLocalEventManager().queueLocalEvent(tfw);
-      }
-    }
-  }
-
-  private updatePatients(state: MainSimulationState, timeJump: number) {
-    const patients = state.getInternalStateObject().patients;
-    computeNewPatientsState(patients, timeJump);
-  }
-
-  private updateActions(state: MainSimulationState) {
-    state.getInternalStateObject().actions.forEach(a => a.update(state));
-  }
-
-  private updateTasks(state: MainSimulationState) {
-    TaskState.getAllTasks(state).forEach(t => t.update(state, this.extensionProps.timeJump));
-  }
-}
-
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-// actors
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-
-export class AddActorLocalEvent extends LocalEventBase {
-  /**
-   * Adds an actor in the game
-   */
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly role: InterventionRole; // spawned role
-      readonly location?: LOCATION_ENUM | undefined; // if undefined automatically resolved
-      readonly travelTime?: SimDuration; // if 0 no travel time, if greater, a travel action is planned
-    }
-  ) {
-    super({ ...props, type: 'AddActorLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    const actor = new Actor(this.props.role);
-    const loc = this.props.location || actor.getComputedSymbolicLocation(state);
-    actor.setLocation(loc);
-    state.getInternalStateObject().actors.push(actor);
-
-    if (this.props.travelTime != undefined && this.props.travelTime > 0) {
-      actor.setLocation(LOCATION_ENUM.remote);
-      const now = state.getSimTime();
-      const travelAction = new OnTheRoadAction(
-        now,
-        this.props.travelTime,
-        'on-the-road',
-        0,
-        actor.Uid,
-        '' // TODO SAM add the template id (from UniqueActionTemplates list -> InternalActionTemplates)
-      );
-      state.getInternalStateObject().actions.push(travelAction);
-    }
-  }
-}
-
-export class MoveActorLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly actorUid: ActorId;
-      readonly location: LOCATION_ENUM;
-    }
-  ) {
-    super({ ...props, type: 'MoveActorLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    const so = state.getInternalStateObject();
-    // TODO Replace with canMoveToLocation2
-    if (!canMoveToLocation(state, 'Actors', this.props.location)) {
-      mainSimLogger.warn('The actor could not be moved as the target location is invalid');
-    } else {
-      so.actors
-        .filter(a => a.Uid === this.props.actorUid)
-        .forEach(a => (a.Location = this.props.location));
-    }
-  }
-}
-
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-// radio
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-
-export class AddMessageLocalEvent extends LocalEventBase {
-  private static RadioIdProvider = 1;
-
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly senderId?: ActorId | undefined;
-      readonly senderName?: string | undefined; // in case there is no sending actor, free text sender name
-      readonly recipientId?: ActorId | undefined;
-      readonly message: TranslationKey;
-      readonly channel?: RadioType | undefined;
-      readonly omitTranslation?: boolean;
-      readonly messageValues?: (string | number)[];
-    }
-  ) {
-    super({ ...props, type: 'AddMessageLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    const msg = this.props.omitTranslation
-      ? this.props.message
-      : getTranslation(
-          'mainSim-actions-tasks',
-          this.props.message,
-          undefined,
-          this.props.messageValues
-        );
-
-    state.getInternalStateObject().radioMessages.push({
-      senderId: this.props.senderId,
-      senderName: this.props.senderName,
-      recipientId: this.props.recipientId,
-      timeStamp: this.props.simTimeStamp,
-      message: msg,
-      uid: AddMessageLocalEvent.RadioIdProvider++,
-      isRadioMessage: this.props.channel != undefined,
-      channel: this.props.channel,
-      pending: false,
-    });
-  }
-}
-
-export class AddRadioMessageLocalEvent extends AddMessageLocalEvent {
-  constructor(
-    readonly extensionProps: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly senderId?: ActorId | undefined;
-      readonly senderName?: string | undefined; // in case there is no sending actor, free text sender name
-      readonly recipientId?: ActorId | undefined;
-      readonly message: TranslationKey;
-      readonly channel: RadioType;
-      readonly omitTranslation?: boolean;
-      readonly messageValues?: (string | number)[];
-    }
-  ) {
-    super({ ...extensionProps });
-  }
-}
-
-export class AddNotificationLocalEvent extends AddMessageLocalEvent {
-  constructor(
-    readonly extensionProps: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly senderId?: ActorId | undefined;
-      readonly senderName?: string | undefined; // in case there is no sending actor, free text sender name
-      readonly recipientId: ActorId;
-      readonly message: TranslationKey;
-      readonly omitTranslation?: boolean;
-      readonly messageValues?: (string | number)[];
-    }
-  ) {
-    super({ ...extensionProps });
   }
 }
 
@@ -691,226 +405,6 @@ export class ResourceArrivalAnnouncementLocalEvent extends LocalEventBase {
 
 // -------------------------------------------------------------------------------------------------
 // -------------------------------------------------------------------------------------------------
-// RESOURCES
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-
-export class ReserveResourcesLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly resourcesId: ResourceId[];
-      readonly actionId: ActionId;
-    }
-  ) {
-    super({ ...props, type: 'ReserveResourcesLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    ResourceState.reserveResources(state, this.props.resourcesId, this.props.actionId);
-  }
-}
-
-export class UnReserveResourcesLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly resourcesId: ResourceId[];
-    }
-  ) {
-    super({ ...props, type: 'UnReserveResourcesLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    ResourceState.unReserveResources(state, this.props.resourcesId);
-  }
-}
-
-abstract class MoveResourcesLocalEventBase extends LocalEventBase {
-  constructor(
-    private readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly type: string;
-      readonly ownerUid: ActorId;
-      readonly targetLocation: LOCATION_ENUM;
-    }
-  ) {
-    super({ ...props });
-  }
-
-  abstract getInvolvedResources(state: MainSimulationState): Resource[];
-
-  applyStateUpdate(state: MainSimulationState): void {
-    // TODO Replace with canMoveToLocation2
-    if (!canMoveToLocation(state, 'Resources', this.props.targetLocation)) {
-      resourceLogger.warn('The resources could not be moved as the target location is invalid');
-      return;
-    }
-
-    const resources = this.getInvolvedResources(state);
-    ResourceState.sendResourcesToLocation(resources, this.props.targetLocation);
-  }
-}
-
-export class MoveResourcesLocalEvent extends MoveResourcesLocalEventBase {
-  constructor(
-    readonly extensionProps: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly ownerUid: ActorId;
-      readonly resourcesId: ResourceId[];
-      readonly targetLocation: LOCATION_ENUM;
-    }
-  ) {
-    super({ ...extensionProps, type: 'MoveResourcesLocalEvent' });
-  }
-
-  override getInvolvedResources(state: MainSimulationState): Resource[] {
-    return this.extensionProps.resourcesId.map(resourceId =>
-      ResourceState.getResourceById(state, resourceId)
-    );
-  }
-}
-
-export class MoveFreeHumanResourcesByLocationLocalEvent extends MoveResourcesLocalEventBase {
-  constructor(
-    readonly extensionProps: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly ownerUid: ActorId;
-      readonly sourceLocation: LOCATION_ENUM;
-      readonly targetLocation: LOCATION_ENUM;
-    }
-  ) {
-    super({
-      ...extensionProps,
-      type: 'MoveFreeHumanResourcesByLocationLocalEvent',
-    });
-  }
-
-  override getInvolvedResources(state: MainSimulationState): Resource[] {
-    return ResourceState.getFreeHumanResourcesByLocation(state, this.extensionProps.sourceLocation);
-  }
-}
-
-export class MoveFreeWaitingResourcesByTypeLocalEvent extends MoveResourcesLocalEventBase {
-  constructor(
-    readonly extensionProps: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly ownerUid: ActorId;
-      readonly resourceType: ResourceType;
-      readonly targetLocation: LOCATION_ENUM;
-    }
-  ) {
-    super({
-      ...extensionProps,
-      type: 'MoveFreeWaitingResourcesByTypeLocalEvent',
-    });
-  }
-
-  override getInvolvedResources(state: MainSimulationState): Resource[] {
-    return ResourceState.getFreeWaitingResourcesByType(state, this.extensionProps.resourceType);
-  }
-}
-
-export class MoveResourcesAtArrivalLocationLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly resourcesIds: ResourceId[];
-    }
-  ) {
-    super({ ...props, type: 'MoveResourcesAtArrivalLocationLocalEvent' });
-  }
-
-  override applyStateUpdate(state: MainSimulationState) {
-    this.props.resourcesIds.forEach(resourceId => {
-      const resource: Resource = ResourceState.getResourceById(state, resourceId);
-      const location = ResourceLogic.resourceArrivalLocationResolution(state, resource.type);
-      ResourceState.sendResourcesToLocation([resource], location);
-    });
-  }
-}
-
-export class AssignResourcesToTaskLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly resourcesId: ResourceId[];
-      readonly taskId: TaskId;
-    }
-  ) {
-    super({ ...props, type: 'AssignResourcesToTaskLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    ResourceState.assignResourcesToTask(state, this.props.resourcesId, this.props.taskId);
-  }
-}
-
-export class AssignResourcesToWaitingTaskLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly resourcesId: ResourceId[];
-    }
-  ) {
-    super({ ...props, type: 'AssignResourcesToWaitingTaskLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    ResourceState.assignResourcesToTask(state, this.props.resourcesId, getIdleTaskUid(state));
-  }
-}
-
-export class ReleaseResourcesFromTaskLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly taskId: TaskId;
-    }
-  ) {
-    super({ ...props, type: 'ReleaseResourcesFromTaskLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    const involvedResources: Resource[] = ResourceState.getFreeResourcesByTask(
-      state,
-      this.props.taskId
-    );
-    const involvedResourcesId: ResourceId[] = involvedResources.map(
-      (resource: Resource) => resource.Uid
-    );
-    const location: LOCATION_ENUM = TaskState.getTaskResponsibleActorSymbolicLocation(
-      state,
-      this.props.taskId
-    );
-
-    ResourceState.assignResourcesToTask(state, involvedResourcesId, getIdleTaskUid(state));
-    ResourceState.sendResourcesToLocation(involvedResources, location);
-  }
-}
-
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
 // TASKS
 // -------------------------------------------------------------------------------------------------
 // -------------------------------------------------------------------------------------------------
@@ -954,235 +448,6 @@ export class MovePatientLocalEvent extends LocalEventBase {
 
   applyStateUpdate(state: MainSimulationState): void {
     changePatientLocation(state, this.props.patientId, this.props.location);
-  }
-}
-
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-// HOSPITAL
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-
-export class HospitalRequestUpdateLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly senderId: ActorId | undefined;
-      readonly hospitalRequestPayload: HospitalRequestPayload;
-    }
-  ) {
-    super({ ...props, type: 'HospitalRequestUpdateLocalEvent' });
-  }
-
-  private formatHospitalResponse(message: HospitalRequestPayload): string {
-    const hospitals = Object.values(getCachedHospitalsByProximity(message.proximity));
-    const units: PatientUnitId[] = getCachedPatientUnitIdsSorted();
-
-    let casuMessage = '';
-    let qty = 0;
-    for (const hospital of hospitals) {
-      casuMessage += `${hospital.shortName}: \n`;
-
-      for (const unitId of units) {
-        qty = hospital.units[unitId] ?? 0;
-        if (qty > 0) {
-          casuMessage += `${qty} ${I18n.translate(getCachedPatientUnitById(unitId).name)} \n`;
-        }
-      }
-
-      casuMessage += '\n';
-    }
-    return casuMessage;
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    updateHospitalProximityRequest(state, this.props.hospitalRequestPayload.proximity);
-    const evt = new AddRadioMessageLocalEvent({
-      parentEventId: this.props.parentEventId,
-      source: this.props.source,
-      simTimeStamp: this.props.simTimeStamp,
-      senderId: getCasuActorId(),
-      recipientId: this.props.senderId,
-      message: this.formatHospitalResponse(this.props.hospitalRequestPayload),
-      channel: RadioType.CASU,
-      omitTranslation: true,
-    });
-    getLocalEventManager().queueLocalEvent(evt);
-  }
-}
-
-/*
-Pretriage Report calculations and radio response
-*/
-export class PretriageReportResponseLocalEvent extends LocalEventBase {
-  private channel: RadioType = RadioType.RESOURCES;
-
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly senderName: string;
-      readonly recipient: number;
-      readonly pretriageLocation: LOCATION_ENUM;
-      readonly feedbackWhenReport: TranslationKey;
-    }
-  ) {
-    super({ ...props, type: 'PretriageReportResponseLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    const taskStatus: TaskStatus = getTaskCurrentStatus(
-      state,
-      getTaskByTypeAndLocation(state, TaskType.Pretriage, this.props.pretriageLocation).Uid
-    );
-
-    getLocalEventManager().queueLocalEvent(
-      new AddRadioMessageLocalEvent({
-        parentEventId: this.props.parentEventId,
-        source: this.props.source,
-        simTimeStamp: this.props.simTimeStamp,
-        senderName: this.props.senderName,
-        recipientId: this.props.recipient,
-        message:
-          taskStatus === 'Uninitialized'
-            ? getTranslation('mainSim-actions-tasks', 'pretriage-task-notStarted', true, [
-                getTranslation('mainSim-locations', 'location-' + this.props.pretriageLocation),
-              ])
-            : formatStandardPretriageReport(
-                state,
-                this.props.pretriageLocation,
-                this.props.feedbackWhenReport,
-                false,
-                true
-              ),
-        channel: this.channel,
-        omitTranslation: true,
-      })
-    );
-  }
-}
-
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-// ACTIVABLE
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
-
-/**
- * Change the active status of an activable
- */
-export class ChangeActivableStatusLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly target: Uid;
-      readonly option: ActivationOperator;
-    }
-  ) {
-    super({ ...props, type: 'ChangeActivableStatusLocalEvent' });
-  }
-
-  applyStateUpdate(state: MainSimulationState): void {
-    const so = state.getInternalStateObject();
-    const target: Activable | undefined = so.activables[this.props.target];
-    if (target != undefined) {
-      if (this.props.option === 'activate') {
-        target.active = true;
-      } else if (this.props.option === 'deactivate') {
-        target.active = false;
-      } else {
-        activableLogger.error('Unhandled option for changing an activable status', this.props);
-      }
-    } else {
-      activableLogger.error('Could not find activable', this.props);
-    }
-  }
-}
-
-export class ChangeMapActivableStatusLocalEvent extends ChangeActivableStatusLocalEvent {
-  constructor(
-    readonly extensionProps: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly target: Uid;
-      readonly option: ActivationOperator;
-    },
-    readonly buildStatus: BuildStatus
-  ) {
-    super({ ...extensionProps });
-  }
-
-  override applyStateUpdate(state: MainSimulationState): void {
-    const so = state.getInternalStateObject();
-    const target: Activable | undefined = so.activables[this.props.target];
-    if (target != undefined && target.activableType === 'mapEntity') {
-      target.buildStatus = this.buildStatus;
-      if (this.props.option === 'activate') {
-        target.active = true;
-      } else if (this.props.option === 'deactivate') {
-        target.active = false;
-      } else {
-        activableLogger.error('Unhandled option for changing an activable status', this.props);
-      }
-    } else {
-      activableLogger.error('Could not find activable', this.props);
-    }
-  }
-}
-
-/**
- * Change the number of times that a trigger / action template / choice what run
- */
-export class IncrementCountLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly target: Uid;
-    }
-  ) {
-    super({ ...props, type: 'IncrementCountLocalEvent' });
-  }
-
-  override applyStateUpdate(state: MainSimulationState): void {
-    const so = state.getInternalStateObject();
-    const target: Activable | undefined = so.activables[this.props.target];
-    if (target != undefined && target.activableType === 'trigger') {
-      target.count += 1;
-    }
-  }
-}
-
-export class SelectChoiceEffectLocalEvent extends LocalEventBase {
-  constructor(
-    readonly props: {
-      readonly parentEventId: GlobalEventId;
-      readonly source: SourceType;
-      readonly simTimeStamp: SimTime;
-      readonly target: Uid;
-      readonly effect: Uid;
-    }
-  ) {
-    super({ ...props, type: 'SelectChoiceEffectLocalEvent' });
-  }
-
-  override applyStateUpdate(state: MainSimulationState): void {
-    const targetActivable: ChoiceActivable | undefined = getChoiceActivable(
-      state,
-      this.props.target
-    );
-    if (targetActivable) {
-      targetActivable.selectedEffect = this.props.effect;
-    } else {
-      activableLogger.error('Could not find activable', this.props.target);
-    }
   }
 }
