+++ _common/gameModel/libs/ClientScript/game/common/tasks/taskBasePorter.ts
@@ -4,7 +4,7 @@ import { getPriorityByCategoryId } from '../../pretri/triage';
 import { SimFlag } from '../actions/actionTemplateBase';
 import { Actor, InterventionRole } from '../actors/actor';
 import { ResourceId, TranslationKey } from '../baseTypes';
-import { AddRadioMessageLocalEvent, MovePatientLocalEvent } from '../localEvents/localEventBase';
+import { MovePatientLocalEvent } from '../localEvents/localEventBase';
 import { getLocalEventManager } from '../localEvents/localEventManager';
 import { RadioType } from '../radio/communicationType';
 import * as RadioLogic from '../radio/radioLogic';
@@ -20,6 +20,7 @@ import {
 import * as TaskState from '../simulationState/taskStateAccess';
 import { PorterSubTask } from './subTask';
 import { TaskBase, TaskType } from './taskBase';
+import { AddRadioMessageLocalEvent } from '../localEvents/localEventBaseRadio';
 
 // -------------------------------------------------------------------------------------------------
 // Brancardage task
