+++ _common/gameModel/libs/ClientScript/game/mainSimulationLogic.ts
@@ -18,11 +18,9 @@ import {
 import { FullEvent, getAllEvents, sendEvent } from './common/events/eventUtils';
 import { getCurrentGameOptions } from './common/gameOptions';
 import {
-  AddNotificationLocalEvent,
-  AddRadioMessageLocalEvent,
   GameOptionsUpdateLocalEvent,
   LocalEventBase,
-  TimeForwardLocalEvent,
+
 } from './common/localEvents/localEventBase';
 import { getLocalEventManager } from './common/localEvents/localEventManager';
 import { MainSimulationState } from './common/simulationState/mainSimulationState';
@@ -35,6 +33,8 @@ import {
 import { loadActionTemplates } from './loaders/actionTemplateLoader';
 import { eraseInitialState, getStartingLocalEvents, shallowState } from './loaders/mainStateLoader';
 import { getOmittedGlobalEvents } from './testing/stateDebug';
+import { TimeForwardLocalEvent } from './common/localEvents/localEventBaseTime';
+import { AddNotificationLocalEvent, AddRadioMessageLocalEvent } from './common/localEvents/localEventBaseRadio';
 
 /* all defined action templates */
 let actionTemplates: Record<ActionTemplateUid, ActionTemplateBase> | undefined;
